+++ javascript/src/components/index.ts
@@ -1,3 +1,4 @@
+export * from "./context_menu"
 export * from "./controls"
 export * from "./pagination"
 export * from "./versions_list"