+++ javascript/src/types.ts
@@ -16,6 +16,7 @@ export type ColumnPicks = Record<Column, boolean>
 export type ViewedList = Version["id"][]
 
 export type Config = {
+  allowRollback: boolean
   columns: ColumnPicks
   setColumns: React.Dispatch<React.SetStateAction<ColumnPicks>>
   viewed: ViewedList
@@ -27,12 +28,19 @@ export interface Version {
   item_type: string
   item_id: string
   whodunnit?: string
-  event: string
+  event: "create" | "update" | "destroy"
   created_at: string
   changeset: Record<string, [unknown, unknown]>
   object: Record<string, unknown>
   item_url?: string
   user_url?: string
+}
+
+export interface VersionContextMenuProps {
+  config: Config
+  coords: null | [number, number]
+  setCoords: Function
+  version: Version
 }
 
 export const stickyStyle = {